Refactor structure for go_app Add migration
