+++ local/go_app_config.yml
@@ -1,0 +1,1 @@
+POSTGRES_URL: postgres://admin:secret@postgres:5432/appdb?sslmode=disable
