+++ src/go_app/assets/assets.go
@@ -1,0 +1,12 @@
+package assets
+
+import "embed"
+
+const (
+	PostgresMigrationDir = "migrations/postgres"
+)
+
+// EmbedMigrations within the go_app binary.
+//
+//go:embed migrations/*
+var EmbedMigrations embed.FS
