+++ src/go_app/internal/users/user.go
@@ -1,0 +1,7 @@
+package users
+
+type User struct {
+	ID    int    `json:"id"`
+	Name  string `json:"name"`
+	Email string `json:"email"`
+}
