+++ src/go_app/server/config.go
@@ -1,0 +1,5 @@
+package server
+
+type Config struct {
+	POSTGRES_URL string
+}
